requirements.txt
docker-py==1.5.0
redis
#lupa
